Agrego failover y restructura para app publica y privada
