+++ src/server/privateApp/privateApp.js
@@ -1,8 +1,8 @@
 const express = require('express');
-const { bind } = require('./routes');
+const { bind } = require('../routes');
 
 const app = express();
-const port = 3000;
+const port = 3001;
 app.use(express.json());
 bind(app);
