+++ src/server/publicApp/publicApp.js
@@ -1,12 +1,9 @@
 const express = require('express');
-const { bind } = require('./routes');
-const { setDownState } = require('./middleware/recoveryMiddleware');
+const { bind } = require('../routes');
 
 const app = express();
 const port = 3000;
 app.use(express.json());
-setDownState(true);
-
 bind(app);
 
 module.exports = app.listen(port);
